+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] - 2022-02-12
+### Changed
+- Sounds are automatically downloaded from SoundPad for TTA supporters (5$/mo+)
+- Drag & Drop sounds from sound pad, including ambience sound
+- Warning for TTA supporters (3$/mo) to inform them about sounds being valid for 24h only
+
 ## [3.4.0] - 2022-02-06
 ### Added
 - Support for WebM and FLAC audio format